AI-Service/requirement.txt
fastapi
uvicorn
pdfplumber
scikit-learn
python-multipart
google-generativeai
python-dotenv